requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-jose
passlib[bcrypt]
python-dotenv
pytest
httpx
